Use weaved's creation time as the lower bound for system time
When generating TLS certificate for the web server, this could happen
early before tlsdated is able to update the system clock and we can
generate the certificate which is valid between years 1970 and 1971.

Use weaved's build date as a lower threshold for system time. If the
system time is earlier than the binary's creation date, use the latter
instead when generating the certificate.

BUG: 23897170
Change-Id: I13420537e482fff4643e0dff9d9f085f3c085e0f