Added GetRequestData() method to simplify HTTP request handlers
Added a utility function libwebserv::GetRequestData() to read
the request body asynchronously into memory and invoke a callback
when all the data is read and available. This greatly simplifies
implementation of some of request handlers that don't work with large
requests and need to read and process the data immediately.

Change-Id: I48588449c77f4362db95fb77a89d382c5ddfaec7